Move listable and locking to new section
Added new conversation settings section "Conversation settings" and
moved the listable and locking settings there.
